fix: storybook paths and moved re-usable chat components out of townhalls domain
