+++ src/domains/Chat/Chatbar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Chatbar';
