+++ src/domains/Questions/QuestionForm/QuestionForm.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Page from 'layout/Page';
 import Component from './QuestionForm';
 
-export default { title: 'Domains/Townhall/Question Form' };
+export default { title: 'Domains/Questions/Question Form' };
 
 export function Basic() {
     return (
